int changed to long for paging (to easier support Redis Lists)

+++ SimpleEventMonitor.Core/EventDataStoreBase.cs
@@ -56,11 +56,11 @@ namespace SimpleEventMonitor.Core
             PublishViaSignalR(simpleEvent);
         }
 
-        public abstract int TotalCount { get; }
+        public abstract long TotalCount { get; }
 
         protected abstract void Persist(SimpleEvent evt);
 
-        public abstract IEnumerable<SimpleEvent> GetEvents(int startIdx = 0, int endIdx = int.MaxValue);
+        public abstract IEnumerable<SimpleEvent> GetEvents(long startIdx = 0, long endIdx = long.MaxValue);
 
         private void PublishViaSignalR(SimpleEvent evt)
         {
